Reorg eth rpc
