+++ client/rpc-core/src/lib.rs
@@ -24,7 +24,12 @@ mod net;
 mod web3;
 
 pub use self::{
-	eth::{EthApi, EthApiServer, EthFilterApi, EthFilterApiServer},
+	eth::{
+		EthBlockApi, EthBlockApiServer, EthClientApi, EthClientApiServer, EthExecuteApi,
+		EthExecuteApiServer, EthFeeApi, EthFeeApiServer, EthFilterApi, EthFilterApiServer,
+		EthMiningApi, EthMiningApiServer, EthStateApi, EthStateApiServer, EthSubmitApi,
+		EthSubmitApiServer, EthTransactionApi, EthTransactionApiServer,
+	},
 	eth_pubsub::{EthPubSubApi, EthPubSubApiServer},
 	net::{NetApi, NetApiServer},
 	web3::{Web3Api, Web3ApiServer},
