+++ client/rpc/src/eth/submit.rs
@@ -1,0 +1,440 @@
+use std::sync::Arc;
+
+use ethereum_types::{H160, H256, U256, U64};
+use futures::future::TryFutureExt;
+use jsonrpc_core::{futures::future, BoxFuture, Result};
+
+use sc_client_api::backend::{Backend, StateBackend, StorageProvider};
+use sc_network::{ExHashT, NetworkService};
+use sc_transaction_pool::{ChainApi, Pool};
+use sc_transaction_pool_api::TransactionPool;
+use sp_api::{ApiExt, ProvideRuntimeApi};
+use sp_block_builder::BlockBuilder as BlockBuilderApi;
+use sp_blockchain::HeaderBackend;
+use sp_runtime::{
+	generic::BlockId,
+	traits::{BlakeTwo256, Block as BlockT},
+	transaction_validity::TransactionSource,
+};
+
+use fc_rpc_core::{
+	types::*, EthClientApi as EthClientApiT, EthFeeApi as EthFeeApiT, EthStateApi as EthStateApiT,
+	EthSubmitApi as EthSubmitApiT,
+};
+use fp_rpc::{ConvertTransaction, ConvertTransactionRuntimeApi, EthereumRuntimeRPCApi};
+
+use crate::{
+	eth::{EthClientApi, EthFeeApi, EthStateApi},
+	internal_err,
+	overrides::OverrideHandle,
+	signer::EthSigner,
+};
+
+pub struct EthSubmitApi<B: BlockT, C, P, CT, BE, H: ExHashT, A: ChainApi> {
+	client: Arc<C>,
+	pool: Arc<P>,
+	convert_transaction: Option<CT>,
+	signers: Arc<Vec<Box<dyn EthSigner>>>,
+
+	client_api: EthClientApi<B, C, BE, H>,
+	state_api: EthStateApi<B, C, BE, P, A>,
+	fee_api: EthFeeApi<B, C, BE>,
+}
+
+impl<B: BlockT, C, P, CT, BE, H: ExHashT, A: ChainApi> EthSubmitApi<B, C, P, CT, BE, H, A> {
+	pub fn new(
+		client: Arc<C>,
+		overrides: Arc<OverrideHandle<B>>,
+		backend: Arc<fc_db::Backend<B>>,
+		pool: Arc<P>,
+		graph: Arc<Pool<A>>,
+		convert_transaction: Option<CT>,
+		network: Arc<NetworkService<B, H>>,
+		signers: Arc<Vec<Box<dyn EthSigner>>>,
+		fee_history_limit: u64,
+		fee_history_cache: FeeHistoryCache,
+	) -> Self {
+		let client_api =
+			EthClientApi::new(client.clone(), overrides.clone(), network, signers.clone());
+		let state_api = EthStateApi::new(
+			client.clone(),
+			overrides.clone(),
+			backend.clone(),
+			pool.clone(),
+			graph.clone(),
+		);
+		let fee_api = EthFeeApi::new(
+			client.clone(),
+			overrides.clone(),
+			backend.clone(),
+			fee_history_limit,
+			fee_history_cache,
+		);
+		Self {
+			client,
+			pool,
+			convert_transaction,
+			signers,
+
+			client_api,
+			state_api,
+			fee_api,
+		}
+	}
+}
+
+impl<B, C, P, CT, BE, H: ExHashT, A> EthSubmitApiT for EthSubmitApi<B, C, P, CT, BE, H, A>
+where
+	B: BlockT<Hash = H256> + Send + Sync + 'static,
+	C: ProvideRuntimeApi<B> + StorageProvider<B, BE>,
+	C: HeaderBackend<B> + Send + Sync + 'static,
+	C::Api: BlockBuilderApi<B> + ConvertTransactionRuntimeApi<B> + EthereumRuntimeRPCApi<B>,
+	BE: Backend<B> + 'static,
+	BE::State: StateBackend<BlakeTwo256>,
+	P: TransactionPool<Block = B> + Send + Sync + 'static,
+	CT: ConvertTransaction<<B as BlockT>::Extrinsic> + Send + Sync + 'static,
+	A: ChainApi<Block = B> + 'static,
+{
+	fn send_transaction(&self, request: TransactionRequest) -> BoxFuture<Result<H256>> {
+		let from = match request.from {
+			Some(from) => from,
+			None => {
+				let accounts = match self.accounts() {
+					Ok(accounts) => accounts,
+					Err(e) => return Box::pin(future::err(e)),
+				};
+
+				match accounts.get(0) {
+					Some(account) => account.clone(),
+					None => return Box::pin(future::err(internal_err("no signer available"))),
+				}
+			}
+		};
+
+		let nonce = match request.nonce {
+			Some(nonce) => nonce,
+			None => match self.transaction_count(from, None) {
+				Ok(nonce) => nonce,
+				Err(e) => return Box::pin(future::err(e)),
+			},
+		};
+
+		let chain_id = match self.chain_id() {
+			Ok(Some(chain_id)) => chain_id.as_u64(),
+			Ok(None) => return Box::pin(future::err(internal_err("chain id not available"))),
+			Err(e) => return Box::pin(future::err(e)),
+		};
+
+		let hash = self.client.info().best_hash;
+
+		let gas_price = request.gas_price;
+		let gas_limit = match request.gas {
+			Some(gas_limit) => gas_limit,
+			None => {
+				let block = self
+					.client
+					.runtime_api()
+					.current_block(&BlockId::Hash(hash));
+				if let Ok(Some(block)) = block {
+					block.header.gas_limit
+				} else {
+					return Box::pin(future::err(internal_err(format!(
+						"block unavailable, cannot query gas limit"
+					))));
+				}
+			}
+		};
+		let max_fee_per_gas = request.max_fee_per_gas;
+		let message: Option<TransactionMessage> = request.into();
+		let message = match message {
+			Some(TransactionMessage::Legacy(mut m)) => {
+				m.nonce = nonce;
+				m.chain_id = Some(chain_id);
+				m.gas_limit = gas_limit;
+				if gas_price.is_none() {
+					m.gas_price = self.gas_price().unwrap_or(U256::default());
+				}
+				TransactionMessage::Legacy(m)
+			}
+			Some(TransactionMessage::EIP2930(mut m)) => {
+				m.nonce = nonce;
+				m.chain_id = chain_id;
+				m.gas_limit = gas_limit;
+				if gas_price.is_none() {
+					m.gas_price = self.gas_price().unwrap_or(U256::default());
+				}
+				TransactionMessage::EIP2930(m)
+			}
+			Some(TransactionMessage::EIP1559(mut m)) => {
+				m.nonce = nonce;
+				m.chain_id = chain_id;
+				m.gas_limit = gas_limit;
+				if max_fee_per_gas.is_none() {
+					m.max_fee_per_gas = self.gas_price().unwrap_or(U256::default());
+				}
+				TransactionMessage::EIP1559(m)
+			}
+			_ => {
+				return Box::pin(future::err(internal_err("invalid transaction parameters")));
+			}
+		};
+
+		let mut transaction = None;
+
+		for signer in &*self.signers {
+			if signer.accounts().contains(&from) {
+				match signer.sign(message, &from) {
+					Ok(t) => transaction = Some(t),
+					Err(e) => return Box::pin(future::err(e)),
+				}
+				break;
+			}
+		}
+
+		let transaction = match transaction {
+			Some(transaction) => transaction,
+			None => return Box::pin(future::err(internal_err("no signer available"))),
+		};
+		let transaction_hash = transaction.hash();
+
+		let block_hash = BlockId::hash(self.client.info().best_hash);
+		let api_version = match self
+			.client
+			.runtime_api()
+			.api_version::<dyn ConvertTransactionRuntimeApi<B>>(&block_hash)
+		{
+			Ok(api_version) => api_version,
+			_ => return Box::pin(future::err(internal_err("cannot access runtime api"))),
+		};
+
+		let extrinsic = match api_version {
+			Some(2) => match self
+				.client
+				.runtime_api()
+				.convert_transaction(&block_hash, transaction)
+			{
+				Ok(extrinsic) => extrinsic,
+				Err(_) => return Box::pin(future::err(internal_err("cannot access runtime api"))),
+			},
+			Some(1) => {
+				if let ethereum::TransactionV2::Legacy(legacy_transaction) = transaction {
+					// To be compatible with runtimes that do not support transactions v2
+					#[allow(deprecated)]
+					match self
+						.client
+						.runtime_api()
+						.convert_transaction_before_version_2(&block_hash, legacy_transaction)
+					{
+						Ok(extrinsic) => extrinsic,
+						Err(_) => {
+							return Box::pin(future::err(internal_err("cannot access runtime api")))
+						}
+					}
+				} else {
+					return Box::pin(future::err(internal_err(
+						"This runtime not support eth transactions v2",
+					)));
+				}
+			}
+			None => {
+				if let Some(ref convert_transaction) = self.convert_transaction {
+					convert_transaction.convert_transaction(transaction.clone())
+				} else {
+					return Box::pin(future::err(internal_err(
+                        "No TransactionConverter is provided and the runtime api ConvertTransactionRuntimeApi is not found"
+                    )));
+				}
+			}
+			_ => {
+				return Box::pin(future::err(internal_err(
+					"ConvertTransactionRuntimeApi version not supported",
+				)))
+			}
+		};
+
+		Box::pin(
+			self.pool
+				.submit_one(&block_hash, TransactionSource::Local, extrinsic)
+				.map_ok(move |_| transaction_hash)
+				.map_err(|err| {
+					internal_err(format!("submit transaction to pool failed: {:?}", err))
+				}),
+		)
+	}
+
+	fn send_raw_transaction(&self, bytes: Bytes) -> BoxFuture<Result<H256>> {
+		let slice = &bytes.0[..];
+		if slice.len() == 0 {
+			return Box::pin(future::err(internal_err("transaction data is empty")));
+		}
+		let first = slice.get(0).unwrap();
+		let transaction = if first > &0x7f {
+			// Legacy transaction. Decode and wrap in envelope.
+			match rlp::decode::<ethereum::TransactionV0>(slice) {
+				Ok(transaction) => ethereum::TransactionV2::Legacy(transaction),
+				Err(_) => return Box::pin(future::err(internal_err("decode transaction failed"))),
+			}
+		} else {
+			// Typed Transaction.
+			// `ethereum` crate decode implementation for `TransactionV2` expects a valid rlp input,
+			// and EIP-1559 breaks that assumption by prepending a version byte.
+			// We re-encode the payload input to get a valid rlp, and the decode implementation will strip
+			// them to check the transaction version byte.
+			let extend = rlp::encode(&slice);
+			match rlp::decode::<ethereum::TransactionV2>(&extend[..]) {
+				Ok(transaction) => transaction,
+				Err(_) => return Box::pin(future::err(internal_err("decode transaction failed"))),
+			}
+		};
+
+		let transaction_hash = transaction.hash();
+
+		let block_hash = BlockId::hash(self.client.info().best_hash);
+		let api_version = match self
+			.client
+			.runtime_api()
+			.api_version::<dyn ConvertTransactionRuntimeApi<B>>(&block_hash)
+		{
+			Ok(api_version) => api_version,
+			_ => return Box::pin(future::err(internal_err("cannot access runtime api"))),
+		};
+
+		let extrinsic = match api_version {
+			Some(2) => match self
+				.client
+				.runtime_api()
+				.convert_transaction(&block_hash, transaction)
+			{
+				Ok(extrinsic) => extrinsic,
+				Err(_) => return Box::pin(future::err(internal_err("cannot access runtime api"))),
+			},
+			Some(1) => {
+				if let ethereum::TransactionV2::Legacy(legacy_transaction) = transaction {
+					// To be compatible with runtimes that do not support transactions v2
+					#[allow(deprecated)]
+					match self
+						.client
+						.runtime_api()
+						.convert_transaction_before_version_2(&block_hash, legacy_transaction)
+					{
+						Ok(extrinsic) => extrinsic,
+						Err(_) => {
+							return Box::pin(future::err(internal_err("cannot access runtime api")))
+						}
+					}
+				} else {
+					return Box::pin(future::err(internal_err(
+						"This runtime not support eth transactions v2",
+					)));
+				}
+			}
+			None => {
+				if let Some(ref convert_transaction) = self.convert_transaction {
+					convert_transaction.convert_transaction(transaction.clone())
+				} else {
+					return Box::pin(future::err(internal_err(
+                        "No TransactionConverter is provided and the runtime api ConvertTransactionRuntimeApi is not found"
+                    )));
+				}
+			}
+			_ => {
+				return Box::pin(future::err(internal_err(
+					"ConvertTransactionRuntimeApi version not supported",
+				)))
+			}
+		};
+
+		Box::pin(
+			self.pool
+				.submit_one(&block_hash, TransactionSource::Local, extrinsic)
+				.map_ok(move |_| transaction_hash)
+				.map_err(|err| {
+					internal_err(format!("submit transaction to pool failed: {:?}", err))
+				}),
+		)
+	}
+}
+
+impl<B, C, P, CT, BE, H: ExHashT, A> EthClientApiT for EthSubmitApi<B, C, P, CT, BE, H, A>
+where
+	B: BlockT<Hash = H256> + Send + Sync + 'static,
+	C: ProvideRuntimeApi<B> + StorageProvider<B, BE>,
+	C: HeaderBackend<B> + Send + Sync + 'static,
+	C::Api: EthereumRuntimeRPCApi<B>,
+	BE: Backend<B> + 'static,
+	BE::State: StateBackend<BlakeTwo256>,
+	P: Send + Sync + 'static,
+	CT: Send + Sync + 'static,
+	A: ChainApi + 'static,
+{
+	fn protocol_version(&self) -> Result<u64> {
+		self.client_api.protocol_version()
+	}
+	fn syncing(&self) -> Result<SyncStatus> {
+		self.client_api.syncing()
+	}
+	fn author(&self) -> Result<H160> {
+		self.client_api.author()
+	}
+	fn accounts(&self) -> Result<Vec<H160>> {
+		self.client_api.accounts()
+	}
+	fn block_number(&self) -> Result<U256> {
+		self.client_api.block_number()
+	}
+	fn chain_id(&self) -> Result<Option<U64>> {
+		self.client_api.chain_id()
+	}
+}
+
+impl<B, C, P, CT, BE, H: ExHashT, A> EthStateApiT for EthSubmitApi<B, C, P, CT, BE, H, A>
+where
+	B: BlockT<Hash = H256> + Send + Sync + 'static,
+	C: ProvideRuntimeApi<B> + StorageProvider<B, BE>,
+	C: HeaderBackend<B> + Send + Sync + 'static,
+	C::Api: BlockBuilderApi<B> + EthereumRuntimeRPCApi<B>,
+	BE: Backend<B> + 'static,
+	BE::State: StateBackend<BlakeTwo256>,
+	P: TransactionPool<Block = B> + Send + Sync + 'static,
+	CT: Send + Sync + 'static,
+	A: ChainApi<Block = B> + 'static,
+{
+	fn balance(&self, address: H160, number: Option<BlockNumber>) -> Result<U256> {
+		self.state_api.balance(address, number)
+	}
+	fn storage_at(&self, address: H160, index: U256, number: Option<BlockNumber>) -> Result<H256> {
+		self.state_api.storage_at(address, index, number)
+	}
+	fn transaction_count(&self, address: H160, number: Option<BlockNumber>) -> Result<U256> {
+		self.state_api.transaction_count(address, number)
+	}
+	fn code_at(&self, address: H160, number: Option<BlockNumber>) -> Result<Bytes> {
+		self.state_api.code_at(address, number)
+	}
+}
+
+impl<B, C, P, CT, BE, H: ExHashT, A> EthFeeApiT for EthSubmitApi<B, C, P, CT, BE, H, A>
+where
+	B: BlockT<Hash = H256> + Send + Sync + 'static,
+	C: ProvideRuntimeApi<B> + StorageProvider<B, BE>,
+	C: HeaderBackend<B> + Send + Sync + 'static,
+	C::Api: EthereumRuntimeRPCApi<B>,
+	BE: Backend<B> + 'static,
+	BE::State: StateBackend<BlakeTwo256>,
+	P: Send + Sync + 'static,
+	CT: Send + Sync + 'static,
+	A: ChainApi<Block = B> + 'static,
+{
+	fn gas_price(&self) -> Result<U256> {
+		self.fee_api.gas_price()
+	}
+	fn fee_history(
+		&self,
+		block_count: U256,
+		newest_block: BlockNumber,
+		reward_percentiles: Option<Vec<f64>>,
+	) -> Result<FeeHistory> {
+		self.fee_api
+			.fee_history(block_count, newest_block, reward_percentiles)
+	}
+}
